move logic of AudioPassThru from 'UIManager' to 'AudioPassThru'
