+++ ReferenceDesign/SampleQTHMI/SDLApps/Templates/AudioPassThru/AudioPassThru.h
@@ -2,7 +2,10 @@
 #define AUDIOPASSTHRU_H
 
 #include <QWidget>
-#include "Common/AppBase.h"
+#include <string>
+#include "SDLApps/Templates/Common/AppBase.h"
+
+using std::string;
 
 class CAudioPassThru : public QWidget {
     Q_OBJECT
@@ -13,13 +16,24 @@ public:
 signals:
 
 public slots:
+    //AudioPassThru popup
+    void OnAudioPassThruFinish();
+    void OnAudioPassThruCancel();
+    void OnAudioPassThruTimeOut();
 
     // QWidget interface
 protected:
     void showEvent(QShowEvent *);
+    void hideEvent(QHideEvent *);
+
+private:
+    //AudioPassThru popup
+    void ShowPopupAudioPassThru(const string &appName, const string &duration, const string &contextA);
+    void HidePopupAudioPassThru();
 
 private:
     AppListInterface *m_pList;
+    bool m_bShowPopUpAudioPassThru;
 };
 
 #endif // AUDIOPASSTHRU_H
