+++ ReferenceDesign/SampleQTHMI/SDLApps/Templates/UIManager.h
@@ -15,7 +15,6 @@ class UIManager : public QWidget, public UIInterface {
 
 public:
     explicit UIManager(QWidget *parent = NULL);
-//    explicit UIManager(AppListInterface *pList, QWidget *parent = NULL);
     ~UIManager();
 
     void SetAppListInterface(AppListInterface *pList);
@@ -47,11 +46,6 @@ public:
 
     int GetCurViewId();
 
-    //AudioPassThru popup
-    void AudioPassThruFinish();
-    void AudioPassThruCancel();
-    void AudioPassThruTimeOut();
-
 signals:
     void onAppShowSignal(int type);
     void onAppActivatedSignal(AppDataInterface *pInterface);
@@ -63,8 +57,6 @@ signals:
 
     void appListUpdateSignal();
 
-    void onAudioPassThruSignal();
-
 public slots:
     void AppShowSlot(int type);
     void OnAppUnregisterSlot(int appId);
@@ -73,23 +65,15 @@ public slots:
     void onVideoStopSlots();
     void loadsdk();
 
-    void onAudioPassThruSlot();
 public:
     SDLAppsView *getSDLAppsView();
 
-private:
-    //AudioPassThru popup
-    void ShowPopupAudioPassThru(const string &appName, const string &duration, const string &contextA);
-    void HidePopupAudioPassThru();
 private:
     int m_iCurUI;
     std::string m_sCurTpln;
     AppListInterface *m_pList;
     QLibrary m_sdk;
     TemplateManager m_TplManager;
-
-    bool m_bShowPopUpAudioPassThru;
-//    void waitMSec(int ms);
 };
 
 #endif // UIMANAGER_H
