+++ ReferenceDesign/SampleQTHMI/SDLApps/app/SDLApps.cpp
@@ -199,15 +199,15 @@ void SDLApps::OnReply(string appId, map<string, string> parameter)
         INFO("[SDLApps] OnReply = %s .", QString::fromStdString( it->second).toStdString().c_str());
         if("AudioPassThruFinish" == it->second)
         {
-            m_pUIManager->AudioPassThruFinish();
+            emit SigAudioPassThruFinish();
         }
         else if("AudioPassThruCancel" == it->second)
         {
-            m_pUIManager->AudioPassThruCancel();
+            emit SigAudioPassThruCancel();
         }
         else if("AudioPassThruTimeOut" == it->second)
         {
-            m_pUIManager->AudioPassThruTimeOut();
+            emit SigAudioPassThruTimeOut();
         }
         else
         {
