+++ ReferenceDesign/SampleQTHMI/SDLApps/app/SDLApps.h
@@ -48,6 +48,10 @@ signals:
     void SigNotify(string appId, map<string, string> parameter);
     void SigReply(string appId, map<string, string> parameter);
 
+    void SigAudioPassThruFinish();
+    void SigAudioPassThruCancel();
+    void SigAudioPassThruTimeOut();
+
 private:
     void StartSDL();
